pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Alexander Lyotov <sippie84@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PySocks = "^1.7.1"
mypy = "^0.910"
wemake-python-styleguide = "^0.15.3"
python-telegram-bot = {extras = ["socks"], version = "^13.7"}
httpx = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
